pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-pipeline"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "chromadb>=0.5",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
rag = "cli:main"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
include = ["rag*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
